+++ breaking-front/src/api/profileEdit.js
@@ -1,13 +1,6 @@
 import { API_PATH } from 'constants/path';
 import api from 'api/api';
 
-export const postProfileData = () => {
-  return api({
-    method: 'post',
-    url: API_PATH.OAUTH2_VALIDATE_JWT,
-  });
-};
-
 export const putProfileEdit = (userData) => {
   return api({
     method: 'put',
